[ALPS04461488] tmem: fix GCC build error
Fix GCC build error in kasan enabled project.

MTK-Commit-Id: 8b22c123dd8e7d885cf47e5db0eeb57bc064dc31

CR-Id: ALPS04461488
Feature: GenieZone

+++ drivers/misc/mediatek/trusted_mem/mtee_devices.c
@@ -185,7 +185,8 @@ static int __init mtee_mchunks_init(void)
 	struct trusted_mem_device *t_device;
 	int idx = 0;
 
-	pr_info("%s:%d (%d)\n", __func__, __LINE__, MTEE_MCHUNKS_DEVICE_COUNT);
+	pr_info("%s:%d (%d)\n", __func__, __LINE__,
+		(int)MTEE_MCHUNKS_DEVICE_COUNT);
 
 	for (idx = 0; idx < MTEE_MCHUNKS_DEVICE_COUNT; idx++) {
 		t_device = create_mtee_mchunk_device(
